+++ app/Http/Controllers/StrategyController.php
@@ -2,7 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Activity;
+use App\Models\ImplementationPlan;
 use App\Models\IntermediateOutcome;
+use App\Models\MajorFinalOutput;
 use App\Models\ProgramAndProject;
 use App\Models\Strategy;
 use Illuminate\Http\Request;
@@ -16,19 +19,35 @@ class StrategyController extends Controller
         //$this->middleware(['auth','verified']);
         $this->model = $model;
     }
-    public function index(Request $request, $id)
+    public function index(Request $request, $id, $ismfo)
     {
         //dd($id);
-        $data = $this->model
+        $data=[];
+
+        if($ismfo==1){
+            //dd("true: ".$ismfo);
+            $data = $this->model
+                ->where('idmfo',$id)
+                ->with('mfos')
+                ->orderBy('created_at', 'desc')
+                ->paginate(10)
+                ->withQueryString();
+        }else{
+            $ismfo=0;
+            //dd("false: ".$ismfo);
+            $data = $this->model
                 ->where('idpaps',$id)
                 ->with('paps')
                 ->orderBy('created_at', 'desc')
                 ->paginate(10)
                 ->withQueryString();
+        }
+
 
         return inertia('Strategies/Index',[
             "data"=>$data,
             "idpaps"=>$id,
+            "ismfo"=>$ismfo,
             "filters" => $request->only(['search']),
             'can'=>[
                 'can_access_validation' => Auth::user()->can('can_access_validation',User::class),
@@ -38,14 +57,24 @@ class StrategyController extends Controller
     }
 
 
-    public function create(Request $request, $id)
+    public function create(Request $request, $id, $ismfo)
     {
-
+        $mfos = MajorFinalOutput::get();
         $paps=ProgramAndProject::get();
-        //dd($paps);
+        $idpaps=0;
+        $idmfo=0;
+        if($ismfo==1){
+            $idmfo = $id;
+        }else{
+            $idpaps = $id;
+        }
+        //dd('$idmfo: '.$idmfo.'idpaps: '.$idpaps);
         return inertia('Strategies/Create',[
-            'idpaps'=>$id,
+            'idpaps'=>$idpaps,
+            'idmfo'=>$idmfo,
             'paps'=>$paps,
+            'mfos'=>$mfos,
+            'ismfo'=>$ismfo,
             'can'=>[
                 'can_access_validation' => Auth::user()->can('can_access_validation',User::class),
                 'can_access_indicators' => Auth::user()->can('can_access_indicators',User::class)
@@ -57,14 +86,25 @@ class StrategyController extends Controller
     public function store(Request $request)
     {
         //dd($request);
+        $ismfo='0';
+
         $attributes = $request->validate([
             'idpaps'=>'required',
+            'idmfo'=>'required',
             'description' => 'required',
         ]);
         $this->model->create($attributes);
         //$request->pass='';
-        return redirect('/strategies/'.$request->idpaps)
+        if($request->idpaps=='0'){
+            $ismfo='1';
+            return redirect('/strategies/'.$request->idmfo.'/'.$ismfo.'/strat/mfo')
                 ->with('message','Outcome added');
+        }else{
+            $ismfo='0';
+            return redirect('/strategies/'.$request->idpaps.'/'.$ismfo.'/strat/mfo')
+                ->with('message','Outcome added');
+        }
+
     }
 
 
@@ -111,9 +151,34 @@ class StrategyController extends Controller
 
     public function destroy(Request $request, $id, $idpaps)
     {
-        $data = $this->model->findOrFail($id);
-        $data->delete();
+        //dd('destroy strat');
+        $strat = $this->model->where('id', $id)->first();
+        //dd($strat->idpaps);
+        $count_imp = ImplementationPlan::where('idstrategy', $id)->count();
+        $count_act = Activity::where('strategy_id', $id)->count();
+        $ismfo="0";
+        if($strat->idpaps=="0"){
+            $ismfo="1";
+        }
+        $msg="";
+        $status ="";
+        if($count_imp>0 || $count_act>0){
+            $msg="Unable to delete!";
+            if($count_imp>0){
+                $msg = $msg." \nDelete its child implementation plans first!";
+            }
+            if($count_act>0){
+                $msg = $msg." \nDelete its child activities first!";
+            }
+            $status ="error";
+        }else{
+            $msg="GAD Issue successfully deleted!";
+            $status ="message";
+
+            $data = $this->model->findOrFail($id);
+            $data->delete();
+        }
         //dd($request->raao_id);
-        return redirect('/strategies/'.$idpaps)->with('warning', 'Strategy deleted');
+        return redirect('/strategies/'.$idpaps.'/'.$ismfo.'/strat/mfo')->with($status, $msg);
     }
 }
